+++ deploy_website.sh
@@ -31,6 +31,7 @@ fi
 # Copy in special files that GitHub wants in the project root.
 cp CHANGELOG.md docs/changelog.md
 cp .github/CONTRIBUTING.md docs/contributing.md
+cp samples/README.md docs/samples.md
 #cp .github/CODE_OF_CONDUCT.md docs/code-of-conduct.md
 
 # Build the site and push the new files up to GitHub