+++ samples/android-app/src/main/kotlin/dev/zacsweers/metro/sample/android/MainActivity.kt
@@ -1,0 +1,35 @@
+// Copyright (C) 2025 Zac Sweers
+// SPDX-License-Identifier: Apache-2.0
+package dev.zacsweers.metro.sample.android
+
+import android.app.Activity
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.FragmentFactory
+import androidx.fragment.app.commit
+import androidx.fragment.app.replace
+import dev.zacsweers.metro.AppScope
+import dev.zacsweers.metro.ContributesIntoMap
+import dev.zacsweers.metro.Inject
+import dev.zacsweers.metro.binding
+
+@ContributesIntoMap(AppScope::class, binding<Activity>())
+@ActivityKey(MainActivity::class)
+@Inject
+class MainActivity(private val fragmentFactory: FragmentFactory) :
+  AppCompatActivity(R.layout.activity_main) {
+
+  init {
+    supportFragmentManager.fragmentFactory = fragmentFactory
+  }
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+    enableEdgeToEdge()
+    super.onCreate(savedInstanceState)
+
+    if (savedInstanceState == null) {
+      supportFragmentManager.commit { replace<CounterFragment>(R.id.fragment_container) }
+    }
+  }
+}