+++ samples/circuit-app/src/commonMain/kotlin/dev/zacsweers/metro/sample/circuit/AppGraph.kt
@@ -1,0 +1,28 @@
+// Copyright (C) 2025 Zac Sweers
+// SPDX-License-Identifier: Apache-2.0
+package dev.zacsweers.metro.sample.circuit
+
+import com.slack.circuit.foundation.Circuit
+import com.slack.circuit.runtime.presenter.Presenter
+import com.slack.circuit.runtime.ui.Ui
+import dev.zacsweers.metro.AppScope
+import dev.zacsweers.metro.DependencyGraph
+import dev.zacsweers.metro.Provides
+import dev.zacsweers.metro.SingleIn
+
+@DependencyGraph(AppScope::class)
+interface AppGraph {
+  val app: CounterAppClass
+
+  @Provides
+  @SingleIn(AppScope::class)
+  fun provideCircuit(
+    presenterFactories: Set<Presenter.Factory>,
+    uiFactories: Set<Ui.Factory>,
+  ): Circuit {
+    return Circuit.Builder()
+      .addPresenterFactories(presenterFactories)
+      .addUiFactories(uiFactories)
+      .build()
+  }
+}